Complete World Shape ownership
In all cases, the World will own all Shapes. This is mostly achieved by
removing the `obj` field from IntersectionComputation, which held an
(unnecessary) reference to objects in the world. The reference prevented
mutable references to the World, which prevented lighting calculations.

An immutable version of `hit` on Intersections is also included in this
commit, although the immutable implementation could use some work.

Tests are yet to be updated.