Remove all notion of a "saved ray"
Basically, a "saved ray" is something that an object keeps after
calculating an intersection (between ray and point).

The textbook only mentions a "saved ray" once, and it makes
intersections somewhat painful, as they force updates on Shapes. With
the current ownership model of the World, trying to mutate objects
within intersection logic doesn't really work.