Handle reflections
A "reflective" attribute is added to the Material record, and the World
implementation is updated to handle reflection color calculations.

Note that a REFLECTION_RECURSION_DEPTH constant is added to make sure
that function "color_at" (and friends) don't recurse indefinitely when
calculating reflections of reflections, e.g. a hall of mirrors.

Some tests are also included.