Put all Shapes in a world behind 'Rc<T>'s
Not ideal, but this allows a bidirectional tree formation for 'Group'
shapes to exist. Moreover, the World doesn't need to fully own all of
its shapes.

The downside is my overreliance on function 'Ref::leak()' which *works*,
but puts a burden on the programmer.

I might come back to this later and remove the 'Ref::leak()' calls,
hopefully using a better paradigm.