Propogate parent transform calculations to children
Most of this logic takes place in function 'propogate_parent_transform'
in file 'shape.rs'.

As of this commit, object-to-world-space transforms are pre-calculated,
preventing any need for a bidirectional tree structure for shapes.
Parent transforms are "propogated" to shape tree leaves--previously,
they were computed on-the-fly.

Future commits will implement these propogations for CSG trees.