Precompute bounds
Shape structures now have a 'saved_bounds' property which prevents
repeated bounds calculations on intersections.

As a note, this uses a 'Cell<T>' on Shape to avoid making every Shape
method requie a mutable reference to Shape (this propogates to almost
every other function requiring intersection).

The downside is that an 'unsafe impl' of trait 'Sync' is required on
'Shape' to pass it between threads--however, the '.set()' operation on
'Cell<T>' should still be safe despite the 'unsafe impl'.